requirements.txt
discord.py==2.6.3
docker==7.1.0

